Move TestClone class
